vm_simulation1/requirements.txt
grpcio
grpcio-tools
flask
colorama
